backend/requirements.txt
aiosmtplib==2.0.1
alembic==1.9.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
bcrypt==4.0.1
billiard==3.6.4.0
celery==5.2.7
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi==1.15.1
charset-normalizer==2.1.1
click @ file:///C:/ci/click_1646056762388/work
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama @ file:///C:/Windows/TEMP/abs_9439aeb1-0254-449a-96f7-33ab5eb17fc8apleb4yn/croots/recipe/colorama_1657009099097/work
cron-descriptor==1.2.35
cryptography==41.0.1
defusedxml==0.7.1
deprecation==2.1.0
dj-database-url==1.3.0
Django==4.1.5
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-cors-headers==3.14.0
django-email-verification==0.3.1
django-filter==23.2
django-heroku==0.3.1
django-rest-framework==0.1.0
django-templated-mail==1.1.1
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
dnspython==2.2.1
email-validator==1.3.0
Flask @ file:///C:/b/abs_ef16l83sif/croot/flask_1671217367534/work
flask-abort==0.0.1
Flask-Bcrypt==1.0.1
Flask-Dance==6.2.0
Flask-Login==0.6.2
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
gunicorn==20.1.0
idna==3.4
itsdangerous @ file:///tmp/build/80754af9/itsdangerous_1621432558163/work
Jinja2 @ file:///C:/b/abs_7cdis66kl9/croot/jinja2_1666908141852/work
kombu==5.2.4
Mako==1.2.4
Markdown==3.4.3
MarkupSafe @ file:///C:/ci/markupsafe_1654508036328/work
node==1.2
nodejs==0.1.1
numpy==1.24.1
oauthlib==3.2.2
odict==1.9.0
optional-django==0.1.0
packaging==23.1
pandas==1.5.3
passlib==1.7.4
Pillow==9.3.0
plumber==1.7
prompt-toolkit==3.0.38
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.7.0
python-crontab==2.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7.1
redis==4.5.4
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
SQLAlchemy==1.4.45
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.13
URLObject==2.4.3
vine==5.0.0
wcwidth==0.2.6
Werkzeug @ file:///C:/b/abs_17q5kgb8bo/croot/werkzeug_1671216014857/work
whitenoise==6.4.0
wincertstore==0.2
WTForms==3.0.1
zope.component==5.1.0
zope.deferredimport==4.4
zope.deprecation==4.4.0
zope.event==4.6
zope.hookable==5.4
zope.interface==5.5.2
zope.lifecycleevent==4.4
zope.proxy==5.0.0
